Add AUTHENTICATION + ERROR-handling

+++ src/main/java/com/adobe/prj/dao/UserDaoJpaImpl.java
@@ -10,7 +10,10 @@ import javax.transaction.Transactional;
 import org.springframework.stereotype.Repository;
 
 import com.adobe.prj.entity.Product;
+import com.adobe.prj.entity.Token;
 import com.adobe.prj.entity.User;
+import com.adobe.prj.exceptions.InvalidAPIKey;
+import com.adobe.prj.exceptions.InvalidParameterOrMissingValue;
 
 @Repository		// All 'Dao' classes should have repository
 public class UserDaoJpaImpl implements UserDao {
@@ -20,12 +23,24 @@ public class UserDaoJpaImpl implements UserDao {
 	private EntityManager em;
 	
 	@Override
-	public User getUser(String email) {
+	public User getUserByEmail(String email) {
 		// Limitation of find is that it can only find based on the primary key
 		// 'find' basically looks at the annotations @Table and @Id to generate the SQL
 		return em.find(User.class, email);
 	}
 
+	@Override
+	public User getUserByAPIKey(String api_key) {
+		String JPQL = "SELECT tok FROM Token tok WHERE tok.apiKey = :api_key";
+		TypedQuery<Token> query = em.createQuery(JPQL, Token.class);
+		query.setParameter("api_key", api_key);
+		List<Token> query_result = query.getResultList();
+		if (query_result.size() <= 0) {
+			throw new InvalidAPIKey("Invalid API Key. Please supply a valid API key");
+		}
+		return query_result.get(0).getUser();
+	}
+	
 	@Override
 	public List<User> getUsers() {
 		String JPQL = "SELECT u FROM User u";
@@ -50,6 +65,9 @@ public class UserDaoJpaImpl implements UserDao {
 	@Transactional
 	public void deleteUser(User u) {
 		User _u = em.find(User.class, u.getEmail());
+		if (_u == null) {
+			throw new InvalidParameterOrMissingValue("Invalid Parameter or Missing Value.");
+		}
 		em.remove(_u);
 	}
